feat: add sent/received timestamps to the metering record
This does not change the engine behavior, but it augments the metering
results with a "timestamps" field (an array of Numbers), which
captures the times at which:
* the primary delivery message was received
* each syscall ("issueCommand" to parent) was sent
* the response from each syscall was received
* the delivery response was sent (actually just before the metering
results are serialized)

The parent process can correlate these timestamps with similar ones
collected on the parent side, to determine how long it took to send
messages over the pipe. If an unexpected stall is observed, this can
help distinguish between the stall happening on the worker process or
on the parent process.

refs https://github.com/Agoric/agoric-sdk/issues/5152