fix: all unknown commands exit with error, add 'q' to quit
Rather than treating all unknown commands as a request for the worker
to silently exit (rc=0), we make a specific 'q' command for that
purpose. Unknown commands now cause an error exit (rc=2), with a
message (to stderr) about what the surprising command was.

This includes the characters that are only used in downstream
responses to upstream `issueCommand()` calls, which might help
diagnose desynchronization errors faster.