add -n command to emit agoric-upgrade string (agoric-upgrade-10)
This is a quick hack to embed a distinctive string into the xsnap
binary, for use by the agoric `agd` command, to ensure that xsnap has
actually been recompiled recently. If a validator somehow forgets to
re-compile xsnap, `agd` will notice the mismatch and complain, rather
than proceeding with the wrong executable.

Running `xsnap -n` emits this "network upgrade version" string,
currently `agoric-upgrade-10`.

In the future, we'll update this string to something new (and perhaps
more fine-grained than a once-per-chain-upgrade value) when we make
changes to xsnap.

refs https://github.com/Agoric/agoric-sdk/issues/7012